[SDK] max.nn.module_v3.Module
The core unit of composition for modeling in MAX.

Informally, a Module is a container class. It can contain other Module
instances, Tensors (the Module's "local parameters") or other arbitrary
Python data.

A Module also has a `__call__` which applies that Module to some input.
In the simplest case this is a function from one Tensor to another
Tensor.

Formally Modules form a tree, and subtrees of Modules can be manipulate
directly. A Module may also be thought of as a closure, where the
parameters form the data of the closure and `__call__` is the
application of the closure.

Terms:
- A "child" of a Module is sub-Module stored directly on that Module.
- A "descendent" of a Module is one of its children, or one of their
descendents.
- A "parameter" is a Tensor storing data on the Module or one of its
descendents.
- The "qualified path" of a descendent is a period-separated string of
the names of the child module attributes which lead to that descendent
module, for instance `child.sub.last`.
- The "qualified path" of a parameter is the qualified path of the
descendent directly holding that parameter, followed by a final path
component for the attribute name of the tensor. For instance `weight`
for a local parameter, or `child.sub.last.weight` for a descendent's
parameter.

```python
>>> from max.experimental.tensor import Tensor
>>> from max.nn.module_v3 import Module, module_dataclass

>>> @module_dataclass
... class Linear(Module):
...     weight: Tensor
...     bias: Tensor | int = 0
...
...     def __call__(self, x: Tensor) -> Tensor:
...     return x @ self.weight.T + self.bias

>>> linear = Linear(Tensor.zeros([5, 4]))
>>> print(linear)
Linear(
  weight=Tensor(shape=[Dim(5), Dim(4)], dtype=DType.bfloat16, device=Device())
)
>>> print(linear(Tensor.constant([1, 2, 3, 4])))
[0., 0., 0., 0., 0.]
```

MAX_INTEGRATION_TESTS_REV_ID: aff34e02e163ba789ac87b7018f6294b32e330b0