[SDK] max.nn.module_v3.Embedding
A vector embedding.

An embedding can be thought of as a lookup table for vectors by index.
Given an input tensor of indices into the embedding, the result of the
embedding lookup is a tensor of the same shape, but with each index
replaced by the value of the vector in that location in the embedding
table.

The common case for embeddings is a 1-dimensional embedding:

```python
>>> from max.dtype import DType
>>> from max.experimental.tensor import Tensor
>>> from max.nn.module_v3 import Embedding

>>> embedding = Embedding(vocab_size=1000, dim=128)
>>> tokens = Tensor.ones([10], dtype=DType.uint64)
>>> embedded = embedding(tokens)
>>> embedded.shape
[Dim(10), Dim(128)]
```

However they just as easily support multi-dimensional embeddings:

```python
>>> from max.dtype import DType
>>> from max.experimental.tensor import Tensor
>>> from max.nn.module_v3 import Embedding

>>> embedding = Embedding(vocab_size=1000, dims=[16, 128])
>>> tokens = Tensor.ones([10], dtype=DType.uint64)
>>> embedded = embedding(tokens)
>>> embedded.shape
[Dim(10), Dim(16), Dim(128)]
```

MAX_INTEGRATION_TESTS_REV_ID: 934b0b7a3786a3f373a4afdc72f6e7dd1da26a96