[SDK] `Module.load_state`
Implements `Module.load_state_dict`, which replaces each parameter in
the module and its descendents.

Also implements `apply_to_local_parameters` and `apply_to_parameters` as
helper functions for `load_state`.

The transformation is applied in-place, updating the module's values and
those of its descendents.

Example:

```python
>>> from max.experimental.tensor import Tensor
>>> from max.nn.module_v3 import Linear

>>> model = Linear(2, 3)
>>> weights = {
...     "weight": Tensor.zeros([3, 2]),
...     "bias": Tensor.zeros([3]),
... }
>>> model.load_state(weights.__getitem__)
```

The lookup is defined as a function rather than a dictionary, allowing
for functional remapping of names during this process to account for
differences in common weight naming and storage conventions.

For instance certain representations may not store weights as
transposed, or may need to be quantized, or split out from a shared qkv
block, or may just have slightly different names or paths.

This can also be used for instance to provide a default value for
initializing LoRA weights.

- The argument to the lookup function is the qualified name of the
parameter with respect to the module on which `load_state` was called.
- The return value of this function is the new value that will replace
the value at that name in the module tree.

Also adds `Module.load_state_dict`:

The transformation is applied in-place, updating the module's values
and those of its descendents.

Example:

```python
>>> from max.experimental.tensor import Tensor
>>> from max.nn.module_v3 import Linear

>>> model = Linear(2, 3)
>>> weights = {
...     "weight": Tensor.zeros([3, 2]),
...     "bias": Tensor.zeros([3]),
... }
>>> model.load_state(weights)
```

Args:
    state: A mapping from qualified name to weight
    strict: If true, verify that every value in `state` is loaded
        at least once.
Raises:
If `strict` is set (default) and not all weights in `state` were loaded.
MAX_INTEGRATION_TESTS_REV_ID: bd241cfbf2b311d1d93f35455c85cab627899fad