[Audio] Move Pipeline._prev_num_steps to Output Field
Holding state on the AudioGenerationPipeline is challenging for a few
reasons:
1. It is difficult to standardize on the Pipeline.execute API, which
presumes that all downstream information surrounding execution is passed
back via the AudioGeneratorOutput class.
2. It is fairly confusing to identify when this field is updated, and
understand its use downstream.

This simplifies this slightly, moving the field to the Output class
itself. This will help with generalizing the scheduler, and remove
unwanted state management within the Pipeline.
This field is identified as an Optional field, as it appears that there
is a variety of downstream usage of the AudioGenerationOutput in the
silence detection and basic processor classes, which do not set and use
this field. While I am not familiar with this code, upon reading it
appears to be quite confusing, we may want to explore if we are
organizing our code correctly here.

MAX_TESTS_TESTS_ORIG_REV_ID: cee74596eb7a483114e54d11f62d5f60a455e5ef