[Pipelines] Generalize Model Specific Required Arguments
This allows the user to pass a required_arguments dictionary in
SupportedArchitecture. Prior to config instantiation we will validate
that the user passes arguments that adhere to these arguments. In the
scenario that these arguments differ, we override the user args to the
required arguments, and print a warning clearly identify this behaviour
to the user.

MAX_INTEGRATION_TESTS_REV_ID: add60c43772fa8627c52587b3447a74e8827a6e3