[Models] Add text-only test for VLMs in smoke tests (#70541)
We've run text-tasks for LLMs and vision-tasks for VLMs for the smoke
tests.
We recently ran into an issue where a VLM had a problem with the
text-only path, but the vision path was OK.

This went undetected in the smoke tests.
This PR makes VLMs run both the vision task and the text-only task, to
ensure this kind of mistake is caught in the future

MAX_INTEGRATION_TESTS_REV_ID: e0d422d4603dce821037e8c1589ccb786f144546