[NIXL] Synchronize before cleanup of transfer engine objects in test to prevent Bad file descriptor error
I got a report that `test_send_recv_concurrent_gpu` is flaky. This test
kicks of 256 concurrent transfers at a single time, greatly stressing
the host.
On the GPU side, this is pretty lightweight since each of the transfers
are
fairly small.

When I run locally, I see:
```
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.569 ms on Device 0 (0.298 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.435 ms on Device 0 (0.299 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.306 ms on Device 0 (0.299 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.150 ms on Device 0 (0.300 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.010 ms on Device 0 (0.300 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.306 ms on Device 0 (0.299 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.150 ms on Device 0 (0.300 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 78.010 ms on Device 0 (0.300 GiB/s)
[SENDER] Transferring 6.00 GB took 107.45 ms (55.84 GB/s)
PASSED
```

On the failing run, I see:
```
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 16680.460 ms on Device 0 (0.001 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 16777.997 ms on Device 0 (0.001 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 16876.611 ms on Device 0 (0.001 GiB/s)
[AsyncRT/lib/NIXL/plugins/ucx/ucx_backend.cpp:325] Transfer of 24.000 MiB took 16975.297 ms on Device 0 (0.001 GiB/s)
[SENDER] Transferring 6.00 GB took 23846.68 ms (0.25 GB/s)
[1761244418.204051] [localhost:232  :0]         eventfd.c:33   UCX  ERROR eventfd error (fd 71 blocking 1): Bad file descriptor
[1761244418.204088] [localhost:232  :0]         eventfd.c:33   UCX  ERROR eventfd error (fd 71 blocking 1): Bad file descriptor
[1761244418.204092] [localhost:232  :0]         eventfd.c:33   UCX  ERROR eventfd error (fd 71 blocking 1): Bad file descriptor
FAILED
```

It seems like the transfers are absolutely crawling when it fails in CI
(.3 GiB/s vs .001 GiB/s). This is 200x slower.

I suspect that the SENDER calls `engine.cleanup()` and invalidates the
file descriptors on the RECEIVER side. The RECEIVER then tried to check
if
the transfer is complete, but this causes a Bad file descriptor error.

My fix is to add some synchronization to ensure both the SENDER and
RECEIVER
are done with all transfers before calling `engine.cleanup()`.

Hopefully this works since I can't really reproduce the flakiness
locally.

I suspect that I could induce it by adding some well placed `time.sleep`
calls,
but I hope this fixes it and I will not have to do this.

MAX_INTEGRATION_TESTS_REV_ID: 1a37313e06e2c600cff0304ec57ce310badcb171