[kernels][gpu] reduce number of loads in reduction for split-k (#71640)
This PR reduces the number of loads for the reduction step in split-k.
The previous version simple decompose the MMA_N into all columns of 16.
This leads to large amount of loads and stores for let's say MMA_N ==
160. This PR adds a heuristic to decompose the MMA_N into least amount
of chunks. For MMA_N == 160, it will decompose to 128 and 32.

MODULAR_ORIG_COMMIT_REV_ID: 558d51f6e1288f55794d60cddbce49add5bb286b