[utils] Change `sync-stubfiles` to a bash script (#71651)
Originally this logic was more complicated and warranted Python, but now
it's mostly in Bazel. This not only simplifies the implementation but
also improves the error messages, before if this failed then you'd get a
useless Python backtrace in addition to the Bazel error.

MODULAR_ORIG_COMMIT_REV_ID: c63773fc1418e938693a4f824a73f03913153584