[SDK] Add axis argument to softmax ops
This patch adds a new `axis` operand to the `softmax` and `logsoftmax`
operations.

The argument is optional and by default it will keep the existing
behavior of using the last axis.

The kernels still need work to support this so at this time only the
last axis can be used.

MODULAR_ORIG_COMMIT_REV_ID: 2519261b68cc67c9f10003fb22aca235130f3a81