[Scheduler] Implement `max-batch-context-length`
For deepseekV3, there is a restriction that the total number of context
cache tokens across all requests in a batch must be less than or equal
to
`max-batch-context-length`. This PR implements this restriction.

Note that this restriction is applied `dp` times (once per replica).

Currently the value of `max-batch-context-length` can be specified via
the CLI flag `max-batch-context-length`. It defaults to `None` in which
case there will be no limit.

In a future PR, we want to make `max-batch-context-length` automagically
determined by a heuristic. This heuristic would need to be implemented
at
the Pipeline Model layer for Deepseek and exposed to MAX Serve during
initialization. We would want to do a calculation based on free memory
like
we do for MemoryEstimation.

MODULAR_ORIG_COMMIT_REV_ID: 18ce96f83f7e7c9bda29785d3102058af3f9b8ce