[Python] Use selectors in BGRecorder
The select system call takes a bitmask of fd's to monitor. This bitmask
is a fixed size. We often run MAX Serve in an environment where we raise
the max fd limit. This means we may have fds that can't be represented
in the bitmask, and hence can't be selected. So Python complains when we
try to select on a big fd like that. The specific error is "ValueError:
filedescriptor out of range in select()".

selectors tries to expose an interface like select, but it uses
different syscalls internally. Usually epoll on Linux and kqueue on BSD
derivatives (incl macOS). These syscalls don't use bitmasks so work fine
with larger fds.

There is some potential overhead in that we're calling more syscalls now
than we were before: e.g. before we'd call select and that's it, now we
call epoll_create, epoll to register, epoll to poll, and close, but this
is not a performance-critical part of the code so this is totally fine,
I think.

MODULAR_ORIG_COMMIT_REV_ID: c7dc12ac2d9f9ffbc3d3dc69e3023073164c3ea9