[kernels] Replace deprecated `handle()` with `borrow()` in MPI module
Replace the deprecated `OwnedDLHandle.handle()` method with the new
`borrow()` method in the MPI shared memory module. The `handle()` method
was deprecated in favor of `borrow()` to better communicate that it
returns a non-owning reference to the library handle.

This change eliminates the last usage of the deprecated `handle()`
method in the codebase. The `handle()` function will be removed soon.

MODULAR_ORIG_COMMIT_REV_ID: c1cb012fbb4acc48f5ac04502f8762f44ded7d21