[AsyncRT] Add a virtual device concept (#71089)
The `VirtualDeviceContext` and `VirtualDevice` can be used for cases
where we want to compile a model, but don't actually need to run it.
Expose this concept to the Python API so we can use it from the
compile-only mode:

[Internal link]

For this PR, we fallback a real device to implement the virtual device.
In future PRs we will remove this fallback and use a separate
compilation-only device abstraction.

This is part of a larger effort to allow cross-compilation of models
which require a GPU to run, but can be compiled on a machine without a
GPU present.

Towards GEX-2483

MODULAR_ORIG_COMMIT_REV_ID: 91a3c9764f9bf68950538e5b3195f709afac4459