[CI] Debrand HF workflow (#72082)
Every time we mention this workflow by the issue ID, Linear buts in and
thinks we're mentioning the issue. Usually we are not actually
referencing this issue. At this point we usually just mean to say it's
items that run in the Hugging Face workflow, so rename to reflect
exactly that.

MAX_INTEGRATION_TESTS_REV_ID: 01e622927bc5108f6769a6f87813bed39c95c386