[mojo-lang] Fix support for linear types that have an explicit del
This clarifies the behavior of CheckLifetimes w.r.t linear types: The
presence
of an `@explicit_destroy` decorator makes a type linear, not the absence
of
a del method. Linear typs can have explicitly declared del's, but
CheckLifetimes
shouldn't call them. This fixes MOCO-2787, and updates a variety of
small
issues with TestSuite which isn't properly handling its linear nature.

This patch also merges a testcase into another to reduce # distinct test
files.

There is a lot of work that can be done to improve linear types support
in the
future. The only reason the testcase gets a del is because it is
inheriting it
from Movable, which doesn't make sense. I filed MOCO-2793 with some
ideas
on how to improve things as we get time to re-eval the design.

MODULAR_ORIG_COMMIT_REV_ID: d3f6eecbe0deaf5a33e3890b05f94e5b407cfc35