[Pipeline] Add pipeline config finalization method
In DeepSeek, we enforce that the `data_parallel_degree` is the same as
the number of devices. Previously, we were updating this parameter in
`PipelineModel.__init__` but it's better to have this done when
initializing the `PipelineConfig`. That way, we don't need to worry
about methods like `estimate_weights_size(pipeline_config)` using
parameters that will be overwritten later.

MODULAR_ORIG_COMMIT_REV_ID: f21b6a1c1030dfcc55ce53d5c518715180a3b4c3