[KVCache] Rename `fetch` to `get_runtime_inputs` and `input_symbols` to `get_symbolic_inputs`
This new name helps convey the true prupose of the `fetch` method. It
simply
returns the kvcache graph inputs for a batch. It does not allocate
blocks
or mutate state within the kvcache.

It is a error if you try to call `get_runtime_inputs` but there is not
enough
blocks allocated to the request.

We also rename `input_symbols` to `get_symbolic_inputs` to match the
above
naming convention. This is done for both kvcache and lora for
consistency.

MAX_TESTS_TESTS_ORIG_REV_ID: afbd86c9e07a314bcf1520c3f79ff37824c7a058