[KVCache] Move memory estimation of number of blocks to KVCacheParams
This streamlines how we do memory estimation for the KVCache. The main
change
is hauling a bunch of the logic from PagedKVCacheManager class to the
KVCacheParams
class.

- I moved the `num_layers` field from the constructor of the
PagedKVCacheManager
class to the KVCacheParams class. This ensures that the KVCacheParams
has
enough info to compute the KVCache tensor shape.
- KVCacheParams now has methods like `.bytes_per_block` and
`.estimated_memory_size`.
- The constructor of PagedKVCacheManager is relatively simpler now.
Instead of
computing the number of blocks there, the `total_num_pages` is just
passed
in via a constructor arg.
- In many tests, we hardcode that the KVCache will have 8 blocks or so
instead
of allocating as many blocks that fits in 1GiB. We may be able to bump
the
gpu mem needed for test down a bit in the future as a result of this.
- Delete paged size from constructor of PagedKVCacheManager since it is
already provided via KVCacheParams.
- A consequence of this is that KVCacheParams is more paged attn
strategy specific.
- We delete many lines of code :)

MAX_TESTS_TESTS_ORIG_REV_ID: a2c644516f7bbbd9ded463b5504080378097541c