[GenericML] Disable a flaky allgather test in Python (#72808)
This is a better way to disable the test than in
[Internal link] This allows mypy to
continue to run on the test and keep it up to date. But the test won't
run on CI, so it's intermittent failures won't impact others.

Towards GEX-2832

MAX_INTEGRATION_TESTS_REV_ID: 3116faa1d02948ba0202e62355fbf8c1d94f84e2