[MAX] Allow passing weights directly to `Module.compile`
The typical pattern in PyTorch for weight loading is to

1. Create a `model`, which generates random weights in memory
2. Load weights as CPU tensors from a file eg. `.safetensors`
3. Call `model.load_state_dict(weights)`, which erases the old model
parameters
4. Call `model.cuda()` to move the weights to accelerator memory
5. Call `model = torch.compile(model)` to trace and compile the model
for inference

This has a few annoying properties
- It initialize unused random weights in step (1)
- (5) doesn't have the freedom to precompute any subgraphs known to
depend on only the weights and then free the now unused weights (what we
call "exec-invariant code")

`Module` today roughly mirrors this pattern:

1. Create a `model` which generates lazy random weights (not in memory)
1. Create a `model`, which generates (lazy, unrealized) random weights
2. Load weights as host tensors
3. Call `model.load_state_dict(weights)` which erases the lazy random
tensors
4. Call `model.to(Accelerator())` to (lazily) move the weights to
accelerator memory
5. Call `compiled = model.compile(*input_types)` to compile the model
for inference

In (5) we actually _do_ execute exec-invariant code, but because the
uncompiled model still exists we don't have a safe way to free those
weights, which can cause an OOM even if the eager and compiled models
independently would it into memory.

Long term we should
- split the compilation and model loading process, giving users the
freedom to free the eager model before loading the compiled model into
memory
- allow the compiled model to "replace" the eager model with the
compiled representation, similar to torch

in the meantime, we can simplify the weight loading process to solve
this problem:

1. Create a `model`, which generates (lazy, unrealized) random weights
2. Load weights as host tensors
3. Call `compiled = model.compile(*input_types, weights=weights)` to
compile the model for inference

This way we never initialize any accelerator memory that isn't used for
the compiled model, allowing a model to be loaded as long as only the
compiled model fits in memory (ie. even if the eager model wouldn't!)

MODULAR_ORIG_COMMIT_REV_ID: df2bd59771b5fe49768dfa1093888d07bbf87412