[Pipelines] Wire-up LoRA shrink QKV and slice optimizations
This PR introduces a handful of optimizations for the LoRA computation.

* We first write out the lora-shrink operation to a tensor of size [3,
M, N] instead of size [M, 3N]. The reshaped output allows us to
eliminate a few tensor slices.
* Our output of the lora-shrink optimization now requires us to manage
how we issue the lora-expand operation in the KV case. We account for
this change in output by extending the offsets and ids of the KV-grouped
gemm so that we can take in an input of size [2M, N]. This also allows
us to eliminate a grouped-matmul.
* Next, we eliminate the weight concating of the LoRA weights that
weren’t getting constant-folded because the LoRA weights are
`mo.constant.external`.
* Finally, we implement a sliced-add that writes the output of the LoRA
up till a specified row index then we write out the remaining unmodified
linear output to the output tensor. This allows us to not write 0’s in
the grouped gemms.

MAX_INTEGRATION_TESTS_REV_ID: 0211e85a824d57a28a03efbb50666b79f8d648be